Remove json dependency from src and create derived NN class
